requirements.txt
# Core ML
scikit-learn>=1.4
numpy>=1.26
pandas>=2.1
joblib>=1.3

# Imbalance handling — REQUIRED for SMOTE
imbalanced-learn>=0.12

# Serving
fastapi>=0.110
uvicorn>=0.29

# Experiment tracking
mlflow>=2.12

# Plotting
matplotlib>=3.8
seaborn>=0.13

# Config
pyyaml>=6.0
